Migrate UpdatesLoggerSpec to Swift Testing to fix flaky PurgeOldLogs test
